+++ interview/tests.py
@@ -34,9 +34,8 @@ from interview.factory import (
 from interview.models import Process, Document, Interview, Offer
 from interview.views import process, minute_edit, minute, interview, close_process, reopen_process
 from pyoupyou.middleware import ExternalCheckMiddleware
-from ref.factory import SubsidiaryFactory, ConsultantFactory
-from ref.models import Consultant, PyouPyouUser
-from ref.models import Consultant, Subsidiary
+from ref.factory import SubsidiaryFactory, PyouPyouUserFactory
+from ref.models import PyouPyouUser, Subsidiary
 
 from django.conf import settings
 from django.utils.translation import activate
@@ -74,14 +73,14 @@ class AccessRestrictionDateTestCase(TestCase):
         self.factory = RequestFactory()
 
         sub = SubsidiaryFactory()
-        self.consultantOld = Consultant.objects.create_consultant("OLD", "old@mail.com", sub, "OLD")
+        self.consultantOld = PyouPyouUser.objects.create_user("OLD", "old@mail.com", "OLD", company=sub)
         sub.responsible = self.consultantOld
         sub.responsible.save()
-        userOld = self.consultantOld.user
+        userOld = self.consultantOld
         userOld.date_joined = datetime.datetime(2016, 1, 1, tzinfo=datetime.timezone.utc)
         userOld.save()
-        self.consultantNew = Consultant.objects.create_consultant("NEW", "new@mail.com", sub, "NEW")
-        userNew = self.consultantNew.user
+        self.consultantNew = PyouPyouUser.objects.create_user("NEW", "new@mail.com", "NEW", company=sub)
+        userNew = self.consultantNew
         userNew.date_joined = datetime.datetime(2017, 11, 1, tzinfo=datetime.timezone.utc)
         userNew.save()
 
@@ -100,25 +99,25 @@ class AccessRestrictionDateTestCase(TestCase):
         middleware.process_request(request)
         request.session.save()
 
-        request.user = self.consultantOld.user
+        request.user = self.consultantOld
         response = process(request, self.p.id)
         self.assertEqual(response.status_code, 200)
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = process(request, self.p.id)
         self.assertEqual(response.status_code, 404)
 
     def test_view_close_process(self):
         request = self.factory.post(reverse("process-close", kwargs={"process_id": self.p.id}))
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = close_process(request, self.p.id)
         self.assertEqual(response.status_code, 404)
 
     def test_view_reopen_process(self):
         request = self.factory.get(reverse("process-reopen", kwargs={"process_id": self.p.id}))
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = reopen_process(request, self.p.id)
         self.assertEqual(response.status_code, 404)
 
@@ -127,7 +126,7 @@ class AccessRestrictionDateTestCase(TestCase):
             reverse("interview-plan", kwargs={"process_id": self.p.id, "interview_id": self.i.id})
         )
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = interview(request, self.p.id, self.i.id, "plan")
         self.assertEqual(response.status_code, 404)
 
@@ -136,7 +135,7 @@ class AccessRestrictionDateTestCase(TestCase):
             reverse("interview-edit", kwargs={"process_id": self.p.id, "interview_id": self.i.id})
         )
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = interview(request, self.p.id, self.i.id, "edit")
         self.assertEqual(response.status_code, 404)
 
@@ -146,11 +145,11 @@ class AccessRestrictionDateTestCase(TestCase):
         middleware.process_request(request)
         request.session.save()
 
-        request.user = self.consultantOld.user
+        request.user = self.consultantOld
         response = minute_edit(request, self.i.id)
         self.assertEqual(response.status_code, 200)
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = minute_edit(request, self.i.id)
         self.assertEqual(response.status_code, 404)
 
@@ -168,11 +167,11 @@ class AccessRestrictionDateTestCase(TestCase):
         middleware.process_request(request)
         request.session.save()
 
-        request.user = self.consultantOld.user
+        request.user = self.consultantOld
         response = minute(request, self.i.id)
         self.assertEqual(response.status_code, 200)
 
-        request.user = self.consultantNew.user
+        request.user = self.consultantNew
         response = minute(request, self.i.id)
         self.assertEqual(response.status_code, 404)
 
@@ -182,8 +181,8 @@ class AccessRestrictionUserTestCase(TestCase):
         self.factory = RequestFactory()
 
         sub = SubsidiaryFactory()
-        self.consultantItw = Consultant.objects.create_consultant("ITW", "itw@mail.com", sub, "ITW")
-        self.consultantRestricted = Consultant.objects.create_consultant("RES", "res@mail.com", sub, "RES")
+        self.consultantItw = PyouPyouUser.objects.create_user("ITW", "itw@mail.com", "ITW", company=sub)
+        self.consultantRestricted = PyouPyouUser.objects.create_user("RES", "res@mail.com", "RES", company=sub)
 
         self.p = ProcessFactory()
         self.i = InterviewFactory(process=self.p)
@@ -195,11 +194,11 @@ class AccessRestrictionUserTestCase(TestCase):
         middleware.process_request(request)
         request.session.save()
 
-        request.user = self.consultantItw.user
+        request.user = self.consultantItw
         response = minute_edit(request, self.i.id)
         self.assertEqual(response.status_code, 200)
 
-        request.user = self.consultantRestricted.user
+        request.user = self.consultantRestricted
         response = minute_edit(request, self.i.id)
         self.assertEqual(response.status_code, 404)
 
@@ -207,10 +206,10 @@ class AccessRestrictionUserTestCase(TestCase):
 class StatusAndNotificationTestCase(TestCase):
     def test_status_and_notification(self):
         subsidiary = SubsidiaryFactory()
-        subsidiaryResponsible = ConsultantFactory(company=subsidiary)
+        subsidiaryResponsible = PyouPyouUserFactory(company=subsidiary)
         subsidiary.responsible = subsidiaryResponsible
         subsidiary.save()
-        interviewer = ConsultantFactory(company=subsidiary)
+        interviewer = PyouPyouUserFactory(company=subsidiary)
 
         # When we create a process
         # Process state will be: WAITING_INTERVIEWER_TO_BE_DESIGNED
@@ -220,7 +219,7 @@ class StatusAndNotificationTestCase(TestCase):
         self.assertEqual(p.state, Process.WAITING_INTERVIEWER_TO_BE_DESIGNED)
         self.assertEqual(list(p.responsible.all()), [subsidiaryResponsible])
         self.assertEqual(len(mail.outbox), 1)
-        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.user.email])
+        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.email])
         mail.outbox = []
 
         # When we create an interview
@@ -235,7 +234,7 @@ class StatusAndNotificationTestCase(TestCase):
         self.assertEqual(i1.state, Interview.WAITING_PLANIFICATION)
         self.assertEqual(list(p.responsible.all()), [interviewer])
         self.assertEqual(len(mail.outbox), 1)
-        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.user.email, interviewer.user.email])
+        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.email, interviewer.email])
         mail.outbox = []
 
         # After interview planification
@@ -251,7 +250,7 @@ class StatusAndNotificationTestCase(TestCase):
 
         self.assertEqual(list(p.responsible.all()), [interviewer])
         self.assertEqual(len(mail.outbox), 1)
-        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.user.email, interviewer.user.email])
+        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.email, interviewer.email])
         mail.outbox = []
 
         # When ITW date is in the past cron will set state to WAIT_INFORMATION for the interview and indirectly to
@@ -276,7 +275,7 @@ class StatusAndNotificationTestCase(TestCase):
         self.assertEqual(i1.state, Interview.GO)
         self.assertEqual(list(p.responsible.all()), [subsidiaryResponsible])
         self.assertEqual(len(mail.outbox), 1)
-        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.user.email])
+        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.email])
         mail.outbox = []
 
         # After we go for a job offer
@@ -289,7 +288,7 @@ class StatusAndNotificationTestCase(TestCase):
         self.assertEqual(Process.objects.get(id=p.id).state, Process.JOB_OFFER)
         self.assertEqual(list(p.responsible.all()), [subsidiaryResponsible])
         self.assertEqual(len(mail.outbox), 1)
-        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.user.email])
+        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.email])
         mail.outbox = []
 
         # After we hired the candidate or we didn't hired him (can be our offer is refused by the candidate for example)
@@ -301,7 +300,7 @@ class StatusAndNotificationTestCase(TestCase):
 
         self.assertEqual(Process.objects.get(id=p.id).state, Process.HIRED)
         self.assertEqual(list(p.responsible.all()), [])
-        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.user.email])
+        self.assertCountEqual(mail.outbox[0].to, [subsidiaryResponsible.email])
         mail.outbox = 0
 
 
@@ -310,8 +309,8 @@ class AnonymizesCanditateTestCase(TestCase):
         self.factory = RequestFactory()
 
         sub = SubsidiaryFactory()
-        self.consultantItw = Consultant.objects.create_consultant("ITW", "itw@mail.com", sub, "ITW")
-        self.consultantRestricted = Consultant.objects.create_consultant("RES", "res@mail.com", sub, "RES")
+        self.consultantItw = PyouPyouUser.objects.create_user("ITW", "itw@mail.com", "ITW", company=sub)
+        self.consultantRestricted = PyouPyouUser.objects.create_user("RES", "res@mail.com", "RES", company=sub)
 
         self.p = ProcessFactory()
 
@@ -385,11 +384,10 @@ class HomeViewTestCase(TestCase):
     def test_dashboard_logged_in(self):
         # create a consultant
         subsidiary = SubsidiaryFactory()
-        consultant = ConsultantFactory(company=subsidiary)
-        user = consultant.user
+        consultant = PyouPyouUserFactory(company=subsidiary)
 
         # log user in
-        self.client.force_login(user=user)
+        self.client.force_login(user=consultant)
 
         # access dashboard
         response = self.client.get(self.url)
@@ -406,11 +404,10 @@ class HomeViewTestCase(TestCase):
     def test_dashboard_no_data(self):
         # create a consultant
         subsidiary = SubsidiaryFactory()
-        consultant = ConsultantFactory(company=subsidiary)
-        user = consultant.user
+        consultant = PyouPyouUserFactory(company=subsidiary)
 
         # log user in
-        self.client.force_login(user=user)
+        self.client.force_login(user=consultant)
 
         # access dashboard
         response = self.client.get(self.url)
@@ -432,8 +429,7 @@ class HomeViewTestCase(TestCase):
     def test_dashboard_with_data(self):
         # create a consultant
         subsidiary = SubsidiaryFactory()
-        consultant = ConsultantFactory(subsidiary=subsidiary.id)
-        user = consultant.user
+        consultant = PyouPyouUserFactory(subsidiary=subsidiary.id)
 
         # create 1 process in actions_needed_processes_table & in related_process_table
         p1 = ProcessFactory()
@@ -454,7 +450,7 @@ class HomeViewTestCase(TestCase):
         ProcessFactory(subsidiary=subsidiary)
 
         # log user in
-        self.client.force_login(user=user)
+        self.client.force_login(user=consultant)
 
         # access dashboard
         response = self.client.get(self.url)
@@ -490,8 +486,7 @@ class ProcessCreationViewTestCase(TestCase):
 
         # create a consultant
         self.subsidiary = SubsidiaryFactory()
-        self.consultant = ConsultantFactory(company=self.subsidiary)
-        self.user = self.consultant.user
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
 
         self.fake = faker.Faker()
 
@@ -503,7 +498,7 @@ class ProcessCreationViewTestCase(TestCase):
 
     def test_process_creation_logged_in(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process creation view
         response = self.client.get(self.url)
@@ -517,7 +512,7 @@ class ProcessCreationViewTestCase(TestCase):
 
     def test_assert_all_needed_forms_are_displayed(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access dashboard
         response = self.client.get(self.url)
@@ -540,7 +535,7 @@ class ProcessCreationViewTestCase(TestCase):
 
     def test_form_submit_with_least_data(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         candidate_name = self.fake.name()
         response = self.client.post(
@@ -563,7 +558,7 @@ class ProcessCreationViewTestCase(TestCase):
 
     def test_form_submit_with_full_data(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         candidate_name = self.fake.name()
         candidate_mail = f"{slugify(candidate_name)}@mail.com"
@@ -610,7 +605,7 @@ class ProcessCreationViewTestCase(TestCase):
 
     def test_correct_display_of_existing_candidate(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # create a new candidate
         candidate_name = self.fake.name()
@@ -653,7 +648,7 @@ class ProcessCreationViewTestCase(TestCase):
 
     def test_form_submit_reusing_candidate(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # create a new candidate
         candidate_name = self.fake.name()
@@ -704,8 +699,7 @@ class ProcessDetailsViewTestCase(TestCase):
     def setUp(self):
         # create a process
         self.subsidiary = SubsidiaryFactory()
-        self.consultant = ConsultantFactory(company=self.subsidiary)
-        self.user = self.consultant.user
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
         self.process = ProcessFactory(
             subsidiary=self.subsidiary,
             contract_type=ContractTypeFactory(),
@@ -726,7 +720,7 @@ class ProcessDetailsViewTestCase(TestCase):
 
     def test_process_details_logged_in(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process details
         response = self.client.get(self.url)
@@ -740,7 +734,7 @@ class ProcessDetailsViewTestCase(TestCase):
 
     def test_process_details_info_display(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process details
         response = self.client.get(self.url)
@@ -797,7 +791,7 @@ class ProcessDetailsViewTestCase(TestCase):
         self.setUp()
 
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # add an interview
         itw1 = InterviewFactory(process=self.process, kind_of_interview=InterviewKindFactory())
@@ -831,8 +825,7 @@ class ProcessDetailsViewTestCase(TestCase):
 class InterviewMinuteViewTestCase(TestCase):
     def setUp(self):
         self.subsidiary = SubsidiaryFactory()
-        self.consultant = ConsultantFactory(company=self.subsidiary)
-        self.user = self.consultant.user
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
         self.process = ProcessFactory(
             subsidiary=self.subsidiary,
             contract_type=ContractTypeFactory(),
@@ -868,7 +861,7 @@ class InterviewMinuteViewTestCase(TestCase):
 
     def test_interview_minute_logged_in(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process creation view
         response = self.client.get(self.url_minute)
@@ -882,7 +875,7 @@ class InterviewMinuteViewTestCase(TestCase):
 
     def test_interview_minute_edit_logged_in(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process creation view
         response = self.client.get(self.url_edit)
@@ -896,7 +889,7 @@ class InterviewMinuteViewTestCase(TestCase):
 
     def test_interview_minute_display(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process creation view
         response = self.client.get(self.url_minute)
@@ -924,7 +917,7 @@ class InterviewMinuteViewTestCase(TestCase):
 
     def test_interview_minute_edit_display(self):
         # log user in
-        self.client.force_login(user=self.user)
+        self.client.force_login(user=self.consultant)
 
         # access process creation view
         response = self.client.get(self.url_edit)
@@ -958,12 +951,12 @@ class PrivilegeLevelTestCase(TestCase):
         self.subsidiary = SubsidiaryFactory()
         self.source = SourcesFactory(category=SourcesCategoryFactory())
 
-        self.consultant = Consultant.objects.create_consultant("TST", "test@mail.com", self.subsidiary, "test")
-        self.consultant.user.date_joined -= relativedelta(
+        self.consultant = PyouPyouUser.objects.create_user("TST", "test@mail.com", "test", company=self.subsidiary)
+        self.consultant.date_joined -= relativedelta(
             years=2
         )  # make sure our consultant can see all the process we create
         self.consultant.limited_to_source = self.source
-        self.consultant.privilege = Consultant.PrivilegeLevel.EXTERNAL_FULL
+        self.consultant.privilege = PyouPyouUser.PrivilegeLevel.EXTERNAL_FULL
         self.consultant.save()
 
         self.offer = OfferFactory(subsidiary=self.subsidiary)
@@ -978,7 +971,7 @@ class PrivilegeLevelTestCase(TestCase):
         )
 
         # log user in
-        self.client.force_login(self.consultant.user)
+        self.client.force_login(self.consultant)
 
     def test_dashboard_display(self):
         response = self.client.get(reverse(views.dashboard))
@@ -1011,12 +1004,11 @@ class PrivilegeLevelTestCase(TestCase):
 class MiddlewareTestCase(TestCase):
     def setUp(self):
         self.subsidiary = SubsidiaryFactory()
-        self.user = PyouPyouUser.objects.create_user("tst")
         # set privilege level but not limited_to_source
-        self.consultant = ConsultantFactory(company=self.subsidiary, user=self.user)
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
 
         self.url = reverse(views.dashboard)
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
 
     def test_access_all_privilege(self):
         # limited_to_source = None; privilege = 1;
@@ -1026,7 +1018,7 @@ class MiddlewareTestCase(TestCase):
 
     def test_access_external_no_source(self):
         # limited_to_source = None; privilege =  2 || 3;
-        self.consultant.privilege = Consultant.PrivilegeLevel.EXTERNAL_FULL
+        self.consultant.privilege = PyouPyouUser.PrivilegeLevel.EXTERNAL_FULL
         self.consultant.save()
 
         response = self.client.get(self.url)
@@ -1035,7 +1027,7 @@ class MiddlewareTestCase(TestCase):
         self.assertEqual(response.status_code, 403)
         self.assertInHTML(needle="Please contact your system administrator", haystack=str(response.content))
 
-        self.consultant.privilege = Consultant.PrivilegeLevel.EXTERNAL_READONLY
+        self.consultant.privilege = PyouPyouUser.PrivilegeLevel.EXTERNAL_READONLY
         self.consultant.save()
 
         response = self.client.get(self.url)
@@ -1047,7 +1039,7 @@ class MiddlewareTestCase(TestCase):
     def test_access_external_and_source(self):
         # limited_to_source != None; privilege = 2 || 3;
         self.consultant.limited_to_source = SourcesFactory(category=SourcesCategoryFactory())
-        self.consultant.privilege = Consultant.PrivilegeLevel.EXTERNAL_FULL
+        self.consultant.privilege = PyouPyouUser.PrivilegeLevel.EXTERNAL_FULL
         self.consultant.save()
 
         response = self.client.get(self.url)
@@ -1058,8 +1050,7 @@ class MiddlewareTestCase(TestCase):
 class InterviewGoalTestCase(TestCase):
     def setUp(self):
         self.subsidiary = SubsidiaryFactory()
-        self.consultant = ConsultantFactory(company=self.subsidiary)
-        self.user = self.consultant.user
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
 
         self.offer = OfferFactory(subsidiary=self.subsidiary)
         self.process = ProcessFactory(subsidiary=self.subsidiary, offer=self.offer)
@@ -1069,7 +1060,7 @@ class InterviewGoalTestCase(TestCase):
             self.interviews.append(InterviewFactory(process=self.process, goal="", next_interview_goal=""))
 
     def test_no_goal_set(self):
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
 
         for itw in self.interviews:
             url = reverse(views.minute, kwargs={"interview_id": itw.id, "slug_info": self.process.candidate.name_slug})
@@ -1091,7 +1082,7 @@ class InterviewGoalTestCase(TestCase):
         self.interviews[0].goal = f"goal for {self.interviews[0].rank}"
         self.interviews[0].save()
 
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
 
         for itw in self.interviews:
             url = reverse(views.minute, kwargs={"interview_id": itw.id, "slug_info": self.process.candidate.name_slug})
@@ -1110,7 +1101,7 @@ class InterviewGoalTestCase(TestCase):
             views.minute, kwargs={"interview_id": self.interviews[-1].id, "slug_info": self.process.candidate.name_slug}
         )
 
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
 
         response = self.client.get(url)
 
@@ -1122,7 +1113,7 @@ class InterviewGoalTestCase(TestCase):
             views.minute, kwargs={"interview_id": self.interviews[1].id, "slug_info": self.process.candidate.name_slug}
         )
 
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
 
         response = self.client.get(url)
 
@@ -1138,19 +1129,19 @@ class OfferFilterGivenSubsidiaryTestCase(TestCase):
 
         self.subsidiary = Subsidiary.objects.first()
 
-        self.consultant = ConsultantFactory(company=self.subsidiary)
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
 
         # create basic data
         for sub in Subsidiary.objects.all():
             for i in range(3):
                 offer = OfferFactory(subsidiary=sub)
 
-                for i in range(5):
+                for j in range(5):
                     ProcessFactory(subsidiary=sub, offer=offer)
 
         ProcessFactory(subsidiary=self.subsidiary, offer=Offer.objects.exclude(subsidiary=self.subsidiary).first())
 
-        self.client.force_login(self.consultant.user)
+        self.client.force_login(self.consultant)
 
     def test_all_offers(self):
         response = self.client.get(reverse(views.offers))
@@ -1184,8 +1175,7 @@ class ImportCognitoFormTestCase(TestCase):
             source_id=self.source.id, sub_id=self.subsidiary.id, prefix=settings.FORM_WEB_HOOK_PREFIX
         )
 
-        self.consultant = ConsultantFactory(company=self.subsidiary)
-        self.user = self.consultant.user
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
 
         OfferFactory(subsidiary=self.subsidiary)  # Id = 1
 
@@ -1194,7 +1184,7 @@ class ImportCognitoFormTestCase(TestCase):
         self.phone = "0606060606"
 
     def test_given_form_full(self):
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
         data = {
             "Form": {
                 "Id": "1",
@@ -1244,7 +1234,7 @@ class ImportCognitoFormTestCase(TestCase):
         self.assertEqual(documents.count(), 2)
 
     def test_given_form_least_data(self):
-        self.client.force_login(self.user)
+        self.client.force_login(self.consultant)
         data = {
             "Content-Type": "application/json",
             "Form": {
@@ -1279,9 +1269,9 @@ class ImportCognitoFormTestCase(TestCase):
 class SeeLinkedProcessCreatedBeforeUserJoinedTestCase(TestCase):
     def setUp(self):
         self.subsidiary = SubsidiaryFactory()
-        self.consultant = ConsultantFactory(company=self.subsidiary)
+        self.consultant = PyouPyouUserFactory(company=self.subsidiary)
 
-        self.client.force_login(self.consultant.user)
+        self.client.force_login(self.consultant)
 
     def test_general_behaviour(self):
         # create some processes and assert that they are correctly displayed
@@ -1303,7 +1293,7 @@ class SeeLinkedProcessCreatedBeforeUserJoinedTestCase(TestCase):
     def test_linked_process_created_before_user_joined_is_displayed(self):
         older_process = ProcessFactory()
         # create a process older that consultant
-        older_process.start_date = self.consultant.user.date_joined - relativedelta(months=1)
+        older_process.start_date = self.consultant.date_joined - relativedelta(months=1)
         older_process.responsible.add(self.consultant)
 
         response = self.client.get(reverse("process-list"))